+++ lambdas/indexer/requirements.txt
@@ -7,13 +7,21 @@
 attrs==24.2.0
     # via jsonschema
 aws-requests-auth==0.4.3
-    # via es_indexer (setup.py)
+    # via quilt-shared
 boto3==1.35.70
-    # via es_indexer (setup.py)
+    # via
+    #   es_indexer (setup.py)
+    #   quilt-shared
+boto3-stubs[s3,sts]==1.38.46
+    # via quilt-shared
 botocore==1.35.70
     # via
     #   boto3
     #   s3transfer
+botocore-stubs==1.38.46
+    # via
+    #   boto3-stubs
+    #   types-aiobotocore
 certifi==2024.8.30
     # via requests
 cffi==1.17.1
@@ -25,7 +33,7 @@ charset-normalizer==3.4.0
 cryptography==44.0.1
     # via pdfminer-six
 elasticsearch==6.8.2
-    # via es_indexer (setup.py)
+    # via quilt-shared
 et-xmlfile==2.0.0
     # via openpyxl
 fcsparser==0.2.8
@@ -38,8 +46,6 @@ jmespath==1.0.1
     # via
     #   boto3
     #   botocore
-jsonpointer==2.4
-    # via es_indexer (setup.py)
 jsonschema==3.2.0
     # via
     #   es_indexer (setup.py)
@@ -49,6 +55,10 @@ jupyter-core==5.7.2
     # via nbformat
 lxml==5.3.0
     # via python-pptx
+mypy-boto3-s3==1.38.44
+    # via boto3-stubs
+mypy-boto3-sts==1.38.38
+    # via boto3-stubs
 nbformat==5.1.3
     # via es_indexer (setup.py)
 numpy==1.26.4
@@ -83,6 +93,8 @@ python-pptx==0.6.23
     # via es_indexer (setup.py)
 pytz==2024.2
     # via pandas
+quilt-shared[boto,es] @ https://github.com/quiltdata/quilt/archive/bc9c170fc936f3110c8ba9249984efb2b8893f2e.zip#subdirectory=py-shared
+    # via es_indexer (setup.py)
 requests==2.32.3
     # via aws-requests-auth
 s3transfer==0.10.4
@@ -94,13 +106,30 @@ six==1.16.0
 strict-rfc3339==0.7
     # via es_indexer (setup.py)
 t4-lambda-shared[mem,preview] @ https://github.com/quiltdata/quilt/archive/d496dffbfb4b7a2ae05f6c1f7f0cb7d5d43bc984.zip#subdirectory=lambdas/shared
-    # via es_indexer (setup.py)
+    # via
+    #   es_indexer (setup.py)
+    #   t4-lambda-shared
 tenacity==9.0.0
     # via es_indexer (setup.py)
 traitlets==5.14.3
     # via
     #   jupyter-core
     #   nbformat
+types-aiobotocore[s3]==2.23.0
+    # via quilt-shared
+types-aiobotocore-s3==2.23.0
+    # via types-aiobotocore
+types-awscrt==0.27.4
+    # via botocore-stubs
+types-s3transfer==0.13.0
+    # via boto3-stubs
+typing-extensions==4.14.0
+    # via
+    #   boto3-stubs
+    #   mypy-boto3-s3
+    #   mypy-boto3-sts
+    #   types-aiobotocore
+    #   types-aiobotocore-s3
 tzdata==2024.2
     # via pandas
 urllib3==2.2.3
